fix pattern issue
The last `.*$` was too greedy, it would allow the last letter to match
and earlier occurrence. For example:

spec/spec_helper.rb when searching for spec would highlight the c
in the first spec, but the spe in the second spec.
With this, it highlights the whole second spec properly.